% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braha.R
\name{asciify}
\alias{asciify}
\title{Convert an UTF-8 string to ascii with UTF-8 characters replaced with
\\uXXXX codes}
\usage{
asciify(x)
}
\arguments{
\item{x}{the string to be converted}
}
\description{
Convert an UTF-8 string to ascii with UTF-8 characters replaced with
\\uXXXX codes
}
