% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braha.R
\name{asciify_file}
\alias{asciify_file}
\title{Convert an UTF-8 file to ascii with UTF-8 characters replaced with
\\uXXXX codes}
\usage{
asciify_file(infile, outfile, suffix = "_X")
}
\arguments{
\item{infile}{the file to be converted}

\item{outfile}{a file to write the converted content to}

\item{suffix}{a character string to add to the file name if no outfile is given}
}
\description{
Convert an UTF-8 file to ascii with UTF-8 characters replaced with
\\uXXXX codes
}
