% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap to imitate sampling distribution for a mean}
\usage{
bootstrap(x, samples = 1000)
}
\arguments{
\item{x}{a vector with the sample to bootstrap over}

\item{samples}{the number of samples to use, default 1000}
}
\value{
a vector of sample means
}
\description{
Bootstrap to imitate sampling distribution for a mean
}
