% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_by_year}
\alias{bootstrap_by_year}
\title{Bootstrap by publication year}
\usage{
bootstrap_by_year(
  data,
  indicator,
  yearcolumn = "Publication_Year",
  samples = 1000
)
}
\arguments{
\item{data}{a data frame including a year column and an indicator column}

\item{indicator}{a string, the name of the indicator column}

\item{yearcolumn}{a string, the name of the year column (default "Publication_Year")}

\item{samples}{the number of samples to generate per year, default 1000}
}
\description{
Run \code{\link{bootstrap}} for an indicator for each publication year
and keep the resulting samples in a data frame with one column per year
}
