% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{bootstrap_graph}
\alias{bootstrap_graph}
\title{Plot boxplots per year for bootstrap samples}
\usage{
bootstrap_graph(samples, ylabel)
}
\arguments{
\item{samples}{a matrix with samples, one named column for each year}

\item{ylabel}{a name for the indicator to use as label}
}
\description{
Plot boxplots per year for bootstrap samples
}
