% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{cf_jcf_plot}
\alias{cf_jcf_plot}
\title{Plot cf and jcf by site with circles proportional to number of publications}
\usage{
cf_jcf_plot(
  df,
  unfilled = c(),
  xintercept = 1,
  yintercept = 1,
  maxsize = 50,
  pal = palette_kth_neo(),
  solid = 0.9,
  linecol = kth_colors("red")
)
}
\arguments{
\item{df}{data frame with indicators by site (needs site, p, cf, jcf)}

\item{unfilled}{vector of sites to draw unfilled circle for (default empty)}

\item{xintercept}{where to put a vertical line, default 1}

\item{yintercept}{where to put a horizontal line, default 1}

\item{maxsize}{maximum size of circle (default 50)}

\item{pal}{colors to use for circles}

\item{solid}{alpha value for circles, 1 = fully solid, 0 = fully transparent}

\item{linecol}{color for horizontal/vertical lines}
}
\description{
Plot cf and jcf by site with circles proportional to number of publications
}
