% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{collaboration_ext}
\alias{collaboration_ext}
\title{Get external collaboration indicators per publication}
\usage{
collaboration_ext(con, uts, orgids)
}
\arguments{
\item{con}{connection to Bibmet}

\item{uts}{list of UT numbers}

\item{orgids}{list of Unified_org_id:s for organisations in SFO}
}
\value{
tibble
}
\description{
Get external collaboration indicators per publication
}
