% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{collaboration_kth}
\alias{collaboration_kth}
\title{Get internal KTH collaboration indicators per publication}
\usage{
collaboration_kth(con, uts)
}
\arguments{
\item{con}{connection to Bibstat}

\item{uts}{list of UTs}
}
\value{
tibble
}
\description{
Get internal KTH collaboration indicators per publication
}
