% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{collaboration_perf_inter}
\alias{collaboration_perf_inter}
\title{Performance indicators, international collaborations}
\usage{
collaboration_perf_inter(data)
}
\arguments{
\item{data}{collaboration status and indicators per publication and site}
}
\description{
Performance indicators, international collaborations
}
