% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{collaboration_perf_natl}
\alias{collaboration_perf_natl}
\title{Performance indicators, national collaborations}
\usage{
collaboration_perf_natl(data)
}
\arguments{
\item{data}{collaboration status and indicators per publication and site}
}
\description{
Performance indicators, national collaborations
}
