% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{collaboration_perf_other}
\alias{collaboration_perf_other}
\title{Performance indicators, other collaborations}
\usage{
collaboration_perf_other(data)
}
\arguments{
\item{data}{collaboration status and indicators per publication and site}
}
\description{
Performance indicators, other collaborations
}
