% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfo-diva.R
\name{diva_url_csv02}
\alias{diva_url_csv02}
\title{Get URL for "csv02" export from DiVA}
\usage{
diva_url_csv02(baseurl, startyear, stopyear)
}
\arguments{
\item{baseurl}{the base url for the DiVA instance}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Get URL for "csv02" export from DiVA
}
