% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfo-diva.R
\name{diva_url_csvall2}
\alias{diva_url_csvall2}
\title{Get URL for "csvall2" export from DiVA}
\usage{
diva_url_csvall2(baseurl, startyear, stopyear)
}
\arguments{
\item{baseurl}{the base url for the DiVA instance}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Get URL for "csvall2" export from DiVA
}
