% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braha.R
\name{doublediff}
\alias{doublediff}
\title{Double sided setdiff}
\usage{
doublediff(set1, set2)
}
\arguments{
\item{set1}{An object that can be treated as a set}

\item{set2}{An object that can be treated as a set}
}
\description{
Double sided setdiff
}
