% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfo-diva.R
\name{fetch_diva_csv02}
\alias{fetch_diva_csv02}
\title{Get "csv02" export from DiVA}
\usage{
fetch_diva_csv02(baseurl, startyear, stopyear, savefile = NULL)
}
\arguments{
\item{baseurl}{the base url for the DiVA instance}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}

\item{savefile}{an optional filename to save the csv in}
}
\description{
Get "csv02" export from DiVA
}
