% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{fundings_ut}
\alias{fundings_ut}
\title{Get fundings per publication}
\usage{
fundings_ut(con, uts)
}
\arguments{
\item{con}{connection to Bibmet}

\item{uts}{list of UT numbers}
}
\value{
tibble
}
\description{
Get fundings per publication
}
