% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab-funding.R
\name{fundings_wash}
\alias{fundings_wash}
\title{Wash funder strings according to rules}
\usage{
fundings_wash(indata, rules)
}
\arguments{
\item{indata}{a data frame with at least a Grant_agency field}

\item{rules}{a set of rules to implement}
}
\value{
tibble
}
\description{
NOTE: At the moment only full string rules are applied!
ToDo:
- Apply all rule types
- Document principles of \code{rules}
}
