% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braha.R
\name{get_most_common}
\alias{get_most_common}
\title{Get the most common identificator for a group
(for example the most common issn for a journal)}
\usage{
get_most_common(data, group, id)
}
\arguments{
\item{data}{the data to use}

\item{group}{the field to group by}

\item{id}{the field to find the most common of}
}
\description{
Get the most common identificator for a group
(for example the most common issn for a journal)
}
