% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators_diva}
\alias{indicators_diva}
\title{Indicators by year from Bibstat for DiVA like publication list}
\usage{
indicators_diva(con, data, startyear, stopyear, cfyear, n_auth = NULL)
}
\arguments{
\item{con}{connection to Bibstat}

\item{data}{publication list}

\item{startyear}{the first publicaiton year to consider}

\item{stopyear}{the last publication year to consider}

\item{cfyear}{the last publication year to count field normalized citation indicators for}

\item{n_auth}{set to a field name with number of unit authors to use fractional counts}
}
\description{
Indicators by year from Bibstat for DiVA like publication list
}
