% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators_ut}
\alias{indicators_ut}
\title{Indicators by year or individually from Bibstat for list of UT numbers}
\usage{
indicators_ut(
  con,
  uts,
  yearly = TRUE,
  startyear,
  stopyear,
  cfyear,
  n_auth = NULL
)
}
\arguments{
\item{con}{connection to Bibstat}

\item{uts}{list of UT numbers}

\item{yearly}{set to true to group by year, otherwise individual indicators}

\item{startyear}{the first publicaiton year to consider}

\item{stopyear}{the last publication year to consider}

\item{cfyear}{the last publication year to use field normalized citation indicators}

\item{n_auth}{set to a field name with number of unit authors to use fractional counts}
}
\description{
Indicators by year or individually from Bibstat for list of UT numbers
}
