% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incicators.R
\name{indics_frac}
\alias{indics_frac}
\title{Citation indicators by year from masterfile like publication list, fractional counts}
\usage{
indics_frac(data, startyear, stopyear, analysisyear)
}
\arguments{
\item{data}{publication list}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}

\item{analysisyear}{the year of analysis, defaults to current year}
}
\description{
Citation indicators by year from masterfile like publication list, fractional counts
}
