% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indics_full_master}
\alias{indics_full_master}
\title{Citation indicators by year from masterfile like publication list, full counts}
\usage{
indics_full_master(data, startyear, stopyear, analysisyear)
}
\arguments{
\item{data}{publication list}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}

\item{analysisyear}{the year of analysis, defaults to current year}
}
\description{
Citation indicators by year from masterfile like publication list, full counts
}
