% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braha.R
\name{intervalify}
\alias{intervalify}
\title{Get intervals for given years and width}
\usage{
intervalify(years, width = 3, from_highest = TRUE)
}
\arguments{
\item{years}{the years to create intervals from}

\item{width}{the width of the intervals to create (default 3)}

\item{from_highest}{if true, start from the highest year, else from lowest (default TRUE)}
}
\description{
Get intervals for given years and width
}
