% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{labelify_edges}
\alias{labelify_edges}
\title{Add edge labels from node labels in Gephiesque network}
\usage{
labelify_edges(network)
}
\arguments{
\item{network}{a named list(nodes, edges)}
}
\value{
a named list(nodes, edges)
}
\description{
Add edge labels from node labels in Gephiesque network
}
