% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{name_address_check}
\alias{name_address_check}
\title{List full name and address for list of UT:s, a last name and an org_id}
\usage{
name_address_check(con, uts, lastname, org_id = 8)
}
\arguments{
\item{con}{a connection to Bibmet}

\item{uts}{list of UT numbers}

\item{lastname}{Last name of the person to check}

\item{org_id}{Unified_org_id of the org of interest (default 8 = KTH)}
}
\description{
List full name and address for list of UT:s, a last name and an org_id
}
