% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_gephi}
\alias{network_gephi}
\title{Network nodes and edges in Gephi friendly format}
\usage{
network_gephi(
  pubs,
  pub_id = "UT",
  label1 = "Name_eng",
  label2 = NULL,
  indicator = NULL
)
}
\arguments{
\item{pubs}{list of publications with one or two columns for node labels}

\item{pub_id}{the column name to use for publication id (default "UT")}

\item{label1}{the column name to use for node labels (default "Name_eng")}

\item{label2}{(optional) a second column name to use for node labels}

\item{indicator}{(optional) a column name to calculate averages per node for}
}
\description{
Network nodes and edges in Gephi friendly format
}
