% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_gephi_separate}
\alias{network_gephi_separate}
\title{Network nodes and edges with specific nodes handled separately}
\usage{
network_gephi_separate(
  pubs,
  separate,
  pub_id = "UT",
  label = "Name_eng",
  indicators = NULL
)
}
\arguments{
\item{pubs}{list of publications with one column for node labels}

\item{separate}{labels to treat separately}

\item{pub_id}{the column name to use for publication id (default "UT")}

\item{label}{the column name to use for node labels (default "Name_eng")}

\item{indicators}{(optional) a list of column names to calculate averages per node/edge for}
}
\description{
The specified nodes will have no edge between them and
publications are only counted towards p if there is a link
to an "external" node (a label outside the separate group)
}
