% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incicators.R
\name{npubs}
\alias{npubs}
\title{Number of publications by year from masterfile like publication list}
\usage{
npubs(data, startyear, stopyear)
}
\arguments{
\item{data}{publication list}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Number of publications by year from masterfile like publication list
}
