% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{npubs_diva}
\alias{npubs_diva}
\title{Number of publications by year from DiVA like publication list}
\usage{
npubs_diva(data, startyear, stopyear)
}
\arguments{
\item{data}{publication list}

\item{startyear}{the first publicaiton year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Number of publications by year from DiVA like publication list
}
