% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilist.R
\name{pi_years}
\alias{pi_years}
\title{Get years by person for PI list}
\usage{
pi_years(pilist, startyear, stopyear)
}
\arguments{
\item{pilist}{data frame with list of PI:s and columns of years 'X2010' etc}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Get years by person for PI list
}
