% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{pmid_from_ut}
\alias{pmid_from_ut}
\title{Get PMID from Bibmet for list of UT}
\usage{
pmid_from_ut(uts, con)
}
\arguments{
\item{uts}{list of UT numbers}

\item{con}{a connection to Bibmet}
}
\description{
Get PMID from Bibmet for list of UT
}
