% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braha.R
\name{recodeNA}
\alias{recodeNA}
\title{Substitute NA with a text string}
\usage{
recodeNA(x, txt)
}
\arguments{
\item{x}{text string that might be NA}

\item{txt}{text string to substitute NA with}
}
\description{
Substitute NA with a text string
}
