% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfo-diva.R
\name{searchauth_diva_csv02}
\alias{searchauth_diva_csv02}
\title{Get "csv02" export from DiVA}
\usage{
searchauth_diva_csv02(baseurl, authors, startyear, stopyear)
}
\arguments{
\item{baseurl}{the base url for the DiVA instance}

\item{authors}{list of authors}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Get "csv02" export from DiVA
}
