% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql_list}
\alias{sql_list}
\title{Make SQL WHERE clause friendly list out of R vector}
\usage{
sql_list(items)
}
\arguments{
\item{items}{vector of items to list (e.g. UT numbers, KTH-ids ...)}
}
\value{
character
}
\description{
Make SQL WHERE clause friendly list out of R vector
}
