% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{stability_by_year}
\alias{stability_by_year}
\title{Stability interval per year}
\usage{
stability_by_year(data, lower = 0.05, upper = 0.95)
}
\arguments{
\item{data}{a matrix with bootstrap samples}

\item{lower}{the lower percentile to return, default 0.05}

\item{upper}{the upper percentile to return, default 0.95}
}
\value{
a list of two quantiles
}
\description{
Stability interval per year
}
