% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{stability_interval}
\alias{stability_interval}
\title{Stability interval from bootstrapping}
\usage{
stability_interval(x, lower = 0.05, upper = 0.95)
}
\arguments{
\item{x}{a vector of sample means}

\item{lower}{the lower percentile to return, default 0.05}

\item{upper}{the upper percentile to return, default 0.95}
}
\value{
a list of two quantiles
}
\description{
Stability interval from bootstrapping
}
