% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{top10_top20_plot}
\alias{top10_top20_plot}
\title{Plot top10 and jcf by site with circles proportional to number of publications}
\usage{
top10_top20_plot(
  df,
  unfilled = c(),
  xintercept = 0.1,
  yintercept = 0.2,
  maxsize = 50,
  pal = palette_kth_neo()
)
}
\arguments{
\item{df}{data frame with indicators by site (needs site, p, top10, top20)}

\item{unfilled}{vector of sites to draw unfilled circle for (default empty)}

\item{xintercept}{where to put a vertical line, default 0.1}

\item{yintercept}{where to put a horizontal line, default 0.2}

\item{maxsize}{maximum size of circle (default 50)}

\item{pal}{colors to use}
}
\description{
Plot top10 and jcf by site with circles proportional to number of publications
}
