% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{topX_plot}
\alias{topX_plot}
\title{Plot top10 and jcf by site with circles proportional to number of publications}
\usage{
topX_plot(
  df,
  unfilled = c(),
  xintercept = 0.1,
  yintercept = 0.2,
  maxsize = 50,
  pal = palette_kth_neo(),
  solid = 0.9
)
}
\arguments{
\item{df}{data frame with indicators by site (needs site, p, top10, top20)}

\item{unfilled}{vector of sites to draw unfilled circle for (default empty)}

\item{xintercept}{top percentage for publications, default 0.1}

\item{yintercept}{top percentage for journals, default 0.2}

\item{maxsize}{maximum size of circle (default 50)}

\item{pal}{colors to use}

\item{solid}{alpha value for circles, 1 = fully solid, 0 = fully transparent}
}
\description{
Plot top10 and jcf by site with circles proportional to number of publications
}
