% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{topXY_plot}
\alias{topXY_plot}
\title{Plot two top indicators by site}
\usage{
topXY_plot(
  df,
  unfilled = c(),
  xintercept = 0.1,
  yintercept = 0.2,
  maxsize = 50,
  pal = palette_kth_neo(),
  solid = 0.9
)
}
\arguments{
\item{df}{data frame with indicators by site}

\item{unfilled}{vector of sites to draw unfilled circle for (default empty)}

\item{xintercept}{top percentage for publications, default 0.1}

\item{yintercept}{top percentage for journals, default 0.2}

\item{maxsize}{maximum size of circle (default 50)}

\item{pal}{colors to use}

\item{solid}{alpha value for circles, 1 = fully solid, 0 = fully transparent}
}
\description{
Plots one circle for each site, with size proportional to number of publications
and X/Y coordinates according to top indicators.
}
\details{
Incoming data frame needs one column \code{site} and two columns \code{topX} and \code{topY}
where \code{X = 100 * xintercept} and \code{Y = 100 * yintercept}.
}
