% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{utlist_sql}
\alias{utlist_sql}
\title{Make SQL WHERE clause friendly list of UT numbers}
\usage{
utlist_sql(uts)
}
\arguments{
\item{uts}{vector of UT numbers (15 chars each)}
}
\value{
character
}
\description{
Make SQL WHERE clause friendly list of UT numbers
}
