% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{xy_plot}
\alias{xy_plot}
\title{Plot indicators X and Y by site with circles proportional to the number of publications}
\usage{
xy_plot(
  df,
  xintercept,
  yintercept,
  xvar,
  yvar,
  xlab,
  ylab,
  graphtitle = NULL,
  sitevar = "site",
  sizevar = "p",
  percentage = FALSE,
  unfilled = c(),
  maxsize = 30,
  pal = palette_kth_neo(18),
  solid = 0.9,
  linecol = kth_colors("red")
)
}
\arguments{
\item{df}{data frame with indicators by site (needs site, p, cf, jcf)}

\item{xintercept}{where to put a vertical line}

\item{yintercept}{where to put a horizontal line}

\item{xvar}{the name of the variable to be plotted on the X axis}

\item{yvar}{the name of the variable to be plotted on the Y axis}

\item{xlab}{the label for the X axis (same as xvar is not given)}

\item{ylab}{the label for the Y axis (same as yvar is not given)}

\item{graphtitle}{main title for graph, default NULL}

\item{sitevar}{the name of the circle label variable, default "site"}

\item{sizevar}{the name of the size variable, defailt "p"}

\item{percentage}{set to TRUE for percentage scales}

\item{unfilled}{vector of sites to draw unfilled circle for, default empty}

\item{maxsize}{maximum size of circle (default 30)}

\item{pal}{colors to use for circles}

\item{solid}{alpha value for circles, 1 = fully solid, 0 = fully transparent}

\item{linecol}{color for horizontal/vertical lines}
}
\description{
Plot indicators X and Y by site with circles proportional to the number of publications
}
