% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{years_boxplot}
\alias{years_boxplot}
\title{Boxplot per year for some indicator with optional reference line}
\usage{
years_boxplot(
  data,
  year = "Publication_Year",
  indicator,
  ylabel = NULL,
  horizontal = NULL,
  perc = FALSE
)
}
\arguments{
\item{data}{a data frame including columns for year and indicator value}

\item{year}{the column holding years, default Publication_Year}

\item{indicator}{the indicator to make boxplots for}

\item{ylabel}{a label for the y axis, default indicator}

\item{horizontal}{height of optional reference line}

\item{perc}{set to TRUE for percent y scale}
}
\description{
Boxplot per year for some indicator with optional reference line
}
