% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_by_year}
\alias{bootstrap_by_year}
\title{Bootstrap by publication year}
\usage{
bootstrap_by_year(data, indicator, year = Publication_year, samples = 1000)
}
\arguments{
\item{data}{a data frame including a year column and an indicator column}

\item{indicator}{the indicator column}

\item{year}{the year column (default "Publication_Year")}

\item{samples}{the number of samples to generate per year, default 1000}
}
\description{
Run \code{\link{bootstrap}} for an indicator for each publication year
and keep the resulting samples in a data frame with one column per year
}
