% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{con_bibmet}
\alias{con_bibmet}
\title{Get connection to Bibmet database server}
\usage{
con_bibmet(
  db = "BIBMET",
  host = Sys.getenv("DBHOST"),
  user = Sys.getenv("DBUSER"),
  pass = Sys.getenv("DBPASS")
)
}
\arguments{
\item{db}{the database to connect to, default BIBMET}

\item{host}{the host database server, default from Sys.getenv}

\item{user}{the database user to connect with, default from Sys.getenv}

\item{pass}{the password for the user to connect, default from Sys.getenv}
}
\description{
Get connection to Bibmet database server
}
