% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfo-diva.R
\name{fetch_diva_csvall2}
\alias{fetch_diva_csvall2}
\title{Get "csvall2" export from DiVA}
\usage{
fetch_diva_csvall2(baseurl, startyear, stopyear)
}
\arguments{
\item{baseurl}{the base url for the DiVA instance}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Get "csvall2" export from DiVA
}
