% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_by_year}
\alias{graph_by_year}
\title{Graph indicator by time}
\usage{
graph_by_year(df, graphvar, timevar = "Publication_Year", horizontal = NULL)
}
\arguments{
\item{df}{data frame with indicators by year}

\item{graphvar}{name of column with indicator to create graph for}

\item{timevar}{name of column with time, default Publication_Year}

\item{horizontal}{where to put a horizontal line, default NULL}
}
\description{
Graph indicator by time
}
