% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators_ut}
\alias{indicators_ut}
\title{Indicators by year or individually from Bibstat for list of UT numbers}
\usage{
indicators_ut(
  con,
  uts,
  individual = FALSE,
  yearly = TRUE,
  cfyear,
  n_auth = NULL
)
}
\arguments{
\item{con}{connection to Bibstat}

\item{uts}{list of UT numbers}

\item{individual}{set to TRUE to get indicators per UT}

\item{yearly}{set to TRUE to group by year}

\item{cfyear}{the last publication year to use field normalized citation indicators}

\item{n_auth}{set to a field name with number of unit authors to use fractional counts}
}
\description{
Indicators by year or individually from Bibstat for list of UT numbers
}
