% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfo-diva.R
\name{searchauth_diva_csvall2}
\alias{searchauth_diva_csvall2}
\title{Get "csvall2" export from DiVA, search by authorId}
\usage{
searchauth_diva_csvall2(baseurl, authors, startyear, stopyear)
}
\arguments{
\item{baseurl}{the base url for the DiVA instance}

\item{authors}{list of authors (Id and/or Orcid)}

\item{startyear}{the first publication year to consider}

\item{stopyear}{the last publication year to consider}
}
\description{
Get "csvall2" export from DiVA, search by authorId
}
